+++ MicroservicioReservas/.idea/workspace.xml
@@ -4,17 +4,36 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="d95ad098-36f9-4704-9e2f-a377190797c1" name="Changes" comment="" />
+    <list default="true" id="1552c5fd-357c-46af-820c-75e1cb388a9d" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/model/dao/ReservaRepository.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/ReservaRestController/ReservaRestController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/Controller/ReservaRestController.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/model/entities/Reserva.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/model/entities/Reserva.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/model/service/IReservaService.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/uceva/microservicioreservas/model/service/IReservaService.java" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
   </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Class" />
+        <option value="Interface" />
+      </list>
+    </option>
+  </component>
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/.." />
   </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
   <component name="PropertiesComponent"><![CDATA[{
   "keyToString": {
+    "node.js.detected.package.eslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
     "node.js.selected.package.tslint": "(autodetect)"
   }
 }]]></component>
