+++ MicroservicioReservas/src/main/java/com/uceva/microservicioreservas/model/service/IReservaService.java
@@ -1,8 +1,8 @@
 package com.uceva.microservicioreservas.model.service;
 
 import java.util.List;
-public interface IReservaService {
+/*public interface IReservaService extends CrudRepository {
 
 
     
-}
+}*/
